backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv
gunicorn
tensorflow
numpy
Pillow
psycopg2-binary
